Patch Adel Gadllah <adel.gadllah@gmx.net> to support files >2GB: Set
2005-07-30  David Sedeño <david@alderia.com>

        Patch Adel Gadllah <adel.gadllah@gmx.net> to support files >2GB:
        * include/gwget_data.h, src/gwget_data.c, src/wget-log.c: Set guint64 instead of
        guint32 to current and total size vars.
        * src/main_window.c: Use file_size string var in gconf instead of
        total_size (gconf can't store int64).
        * src/main_window_cb.c: Cast to guint32.

+++ include/gwget_data.h
@@ -82,12 +82,12 @@ typedef struct
     gboolean use_proxy;         /* Used to use proxy */
     gboolean use_auto_dl;       /* Used to use auto download */
     gchar *line;                /* Used to process the wget output */
-    guint32 total_size;         /* Total file size in bytes */
+    guint64 total_size;         /* Total file size in bytes */
     guint32 total_time;         /* Total time spent in seconds */
     time_t session_start_time;  /* Time at start of this download session */
-    guint32 session_start_size; /* Size at start of this download session */
+    guint64 session_start_size; /* Size at start of this download session */
     guint32 session_elapsed;    /* Time spent in seconds on this session */
-    guint32 cur_size;           /* Current downloaded file size */
+    guint64 cur_size;           /* Current downloaded file size */
     DlState state;              /* State of the download */
     GtkTreeIter file_list;       /* GtkTreeIter where this file is inserted */
     guint id;                   /* File data id */
@@ -111,7 +111,7 @@ void gwget_data_set_filename(GwgetData* gwgetdata,gchar *str);
 void gwget_data_start_download(GwgetData *gwgetdata);
 void gwget_data_set_state (GwgetData *gwgetdata, DlState state);
 void gwget_data_update_statistics (GwgetData *gwgetdata);
-void gwget_data_set_total_size (GwgetData *gwgetdata,guint32 total_size);
+void gwget_data_set_total_size (GwgetData *gwgetdata,guint64 total_size);
 GwgetData* gwget_data_get_selected(void);
 void gwget_data_free(gpointer data);
 void gwget_data_stop_download(GwgetData *data);
