+++ src/main_window_cb.c
@@ -200,14 +200,14 @@ new_download(GwgetData* gwgetdata)
 	gdouble perc;
 	
 	gtk_list_store_append (GTK_LIST_STORE(model), &iter); 
-	size = g_strdup_printf ("%d kB", (gwgetdata->cur_size + 512) / 1024);
+	size = g_strdup_printf ("%d kB", (guint32)(gwgetdata->cur_size + 512) / 1024);
 	gtk_list_store_set (GTK_LIST_STORE(model), &iter,URL_COLUMN,gwgetdata->url,
 						CURRENTSIZE_COLUMN, size, 
 						FILENAME_COLUMN, gwgetdata->filename,
 	    				-1);
 	
 	
-	size = g_strdup_printf ("%d kB", (gwgetdata->total_size + 512) / 1024);
+	size = g_strdup_printf ("%d kB", (guint32)(gwgetdata->total_size + 512) / 1024);
 	gtk_list_store_set (GTK_LIST_STORE(model), &iter, TOTALSIZE_COLUMN, size, -1);
 	
 	perc = gwgetdata->total_size==0?0:((gdouble)gwgetdata->cur_size*100)/(gdouble)gwgetdata->total_size ;
